requirements.txt
pyautogui
keyboard
pillow
pystray
opencv-python
numpy